backend/requirements.txt
Flask
flask-cors
python-dotenv
openai==0.28.0
